% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_figshare_download_url}
\alias{build_figshare_download_url}
\title{Build Figshare download URL for a collection}
\usage{
build_figshare_download_url(collection_id, private_link_id)
}
\arguments{
\item{collection_id}{An integer for the collection identifier.}

\item{private_link_id}{An identifier value for the private link to the
collection.}
}
\value{
A download URL for the specified private Figshare collection.
}
\description{
Build Figshare download URL for a collection
}
\examples{
\dontrun{
  build_figshare_download_url(
    collection_id = 25370686, 
    private_link_id = "58527668245cb63f14f5"
  )
}

}
\keyword{internal}
