% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsumerRequest.R
\name{ConsumerRequest}
\alias{ConsumerRequest}
\title{Create the Consumer Request object}
\description{
This is used as a DSL representing a request
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConsumerRequest$new()}}
\item \href{#method-getMethod}{\code{ConsumerRequest$getMethod()}}
\item \href{#method-setMethod}{\code{ConsumerRequest$setMethod()}}
\item \href{#method-getPath}{\code{ConsumerRequest$getPath()}}
\item \href{#method-setPath}{\code{ConsumerRequest$setPath()}}
\item \href{#method-getQuery}{\code{ConsumerRequest$getQuery()}}
\item \href{#method-setQuery}{\code{ConsumerRequest$setQuery()}}
\item \href{#method-getHeaders}{\code{ConsumerRequest$getHeaders()}}
\item \href{#method-setHeaders}{\code{ConsumerRequest$setHeaders()}}
\item \href{#method-getBody}{\code{ConsumerRequest$getBody()}}
\item \href{#method-setBody}{\code{ConsumerRequest$setBody()}}
\item \href{#method-jsonReady}{\code{ConsumerRequest$jsonReady()}}
\item \href{#method-clone}{\code{ConsumerRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$new(path = "/", method = "GET")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMethod"></a>}}
\if{latex}{\out{\hypertarget{method-getMethod}{}}}
\subsection{Method \code{getMethod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$getMethod()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setMethod"></a>}}
\if{latex}{\out{\hypertarget{method-setMethod}{}}}
\subsection{Method \code{setMethod()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$setMethod(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPath"></a>}}
\if{latex}{\out{\hypertarget{method-getPath}{}}}
\subsection{Method \code{getPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$getPath()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPath"></a>}}
\if{latex}{\out{\hypertarget{method-setPath}{}}}
\subsection{Method \code{setPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$setPath(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getQuery"></a>}}
\if{latex}{\out{\hypertarget{method-getQuery}{}}}
\subsection{Method \code{getQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$getQuery()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setQuery"></a>}}
\if{latex}{\out{\hypertarget{method-setQuery}{}}}
\subsection{Method \code{setQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$setQuery(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-getHeaders}{}}}
\subsection{Method \code{getHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$getHeaders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-setHeaders}{}}}
\subsection{Method \code{setHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$setHeaders(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBody"></a>}}
\if{latex}{\out{\hypertarget{method-getBody}{}}}
\subsection{Method \code{getBody()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$getBody()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBody"></a>}}
\if{latex}{\out{\hypertarget{method-setBody}{}}}
\subsection{Method \code{setBody()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$setBody(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jsonReady"></a>}}
\if{latex}{\out{\hypertarget{method-jsonReady}{}}}
\subsection{Method \code{jsonReady()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$jsonReady()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConsumerRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
