% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_mismatch}
\alias{detect_mismatch}
\title{Detect one-to-one mismatch in variables}
\usage{
detect_mismatch(x, y)
}
\arguments{
\item{x}{A vector of values from the Pandemic PACT dataset.}

\item{y}{A vector of values from the Pandemic PACT dataset to match \code{x}
against.}
}
\value{
A logical vector indicating whether \code{x} and \code{y} have matching
values (TRUE) or not (FALSE).
}
\description{
Detect one-to-one mismatch in variables
}
\examples{
\dontrun{
  detect_mismatch(ResearchInstitutionCountry, ResearchInstitutionRegion)
} 

}
