% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mpox_priority_who}
\alias{get_mpox_priority_who}
\alias{get_mpox_priority_global}
\title{Get Pandemic PACT Mpox priority - WHO Research Priority}
\usage{
get_mpox_priority_who(pact_data)

get_mpox_priority_global(pact_data)
}
\arguments{
\item{pact_data}{A data.frame for the Pandemic PACT dataset read from the
Pandemic PACT website. This is usually obtained via a call to
\code{pact_read_website()}.}
}
\value{
A data.frame of same structure as \code{pact_data} but with
\code{WHOMpoxResearchPriorities} and \code{WHOMpoxResearchSubPriorities} variables
processed/cleaned.
}
\description{
Get Pandemic PACT Mpox priority - WHO Research Priority
}
\examples{
\dontrun{
  get_mpox_who_priority(pact_data)
}

}
\keyword{internal}
