% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_research_categories}
\alias{get_research_categories}
\title{Get Pandemic PACT research category}
\usage{
get_research_categories(pact_data)
}
\arguments{
\item{pact_data}{A data.frame for the Pandemic PACT dataset read from the
Pandemic PACT website. This is usually obtained via a call to
\code{pact_read_website()}.}
}
\value{
A data.frame of same structure as \code{pact_data} but with ResearchCat
and ResearchSubcat variables processed/cleaned.
}
\description{
Get Pandemic PACT research category
}
\examples{
\dontrun{
  get_research_categories(pact_data)
}

}
\keyword{internal}
