% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallerLinux.R
\name{InstallerLinux}
\alias{InstallerLinux}
\title{Class to download and install Ruby}
\description{
Class to download and install Ruby

Class to download and install Ruby
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-isEligible}{\code{InstallerLinux$isEligible()}}
\item \href{#method-install}{\code{InstallerLinux$install()}}
\item \href{#method-clone}{\code{InstallerLinux$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isEligible"></a>}}
\if{latex}{\out{\hypertarget{method-isEligible}{}}}
\subsection{Method \code{isEligible()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstallerLinux$isEligible()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-install"></a>}}
\if{latex}{\out{\hypertarget{method-install}{}}}
\subsection{Method \code{install()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstallerLinux$install(destinationDir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstallerLinux$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
