% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallManager.R
\name{InstallManager}
\alias{InstallManager}
\title{Managing classes to install Ruby}
\description{
Managing classes to install Ruby

Managing classes to install Ruby
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-install}{\code{InstallManager$install()}}
\item \href{#method-setDestinationDir}{\code{InstallManager$setDestinationDir()}}
\item \href{#method-uninstall}{\code{InstallManager$uninstall()}}
\item \href{#method-clone}{\code{InstallManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-install"></a>}}
\if{latex}{\out{\hypertarget{method-install}{}}}
\subsection{Method \code{install()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstallManager$install()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDestinationDir"></a>}}
\if{latex}{\out{\hypertarget{method-setDestinationDir}{}}}
\subsection{Method \code{setDestinationDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstallManager$setDestinationDir(dir)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uninstall"></a>}}
\if{latex}{\out{\hypertarget{method-uninstall}{}}}
\subsection{Method \code{uninstall()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstallManager$uninstall()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InstallManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
