% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interaction.R
\name{Interaction}
\alias{Interaction}
\title{Create the Pact Interaction object}
\description{
This is used as a DSL representing an Interaction
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Interaction$new()}}
\item \href{#method-getDescription}{\code{Interaction$getDescription()}}
\item \href{#method-setDescription}{\code{Interaction$setDescription()}}
\item \href{#method-getProviderState}{\code{Interaction$getProviderState()}}
\item \href{#method-setProviderState}{\code{Interaction$setProviderState()}}
\item \href{#method-getRequest}{\code{Interaction$getRequest()}}
\item \href{#method-setRequest}{\code{Interaction$setRequest()}}
\item \href{#method-getResponse}{\code{Interaction$getResponse()}}
\item \href{#method-setResponse}{\code{Interaction$setResponse()}}
\item \href{#method-jsonReady}{\code{Interaction$jsonReady()}}
\item \href{#method-toJSON}{\code{Interaction$toJSON()}}
\item \href{#method-clone}{\code{Interaction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDescription"></a>}}
\if{latex}{\out{\hypertarget{method-getDescription}{}}}
\subsection{Method \code{getDescription()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$getDescription()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDescription"></a>}}
\if{latex}{\out{\hypertarget{method-setDescription}{}}}
\subsection{Method \code{setDescription()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$setDescription(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProviderState"></a>}}
\if{latex}{\out{\hypertarget{method-getProviderState}{}}}
\subsection{Method \code{getProviderState()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$getProviderState()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProviderState"></a>}}
\if{latex}{\out{\hypertarget{method-setProviderState}{}}}
\subsection{Method \code{setProviderState()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$setProviderState(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRequest"></a>}}
\if{latex}{\out{\hypertarget{method-getRequest}{}}}
\subsection{Method \code{getRequest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$getRequest()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRequest"></a>}}
\if{latex}{\out{\hypertarget{method-setRequest}{}}}
\subsection{Method \code{setRequest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$setRequest(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getResponse"></a>}}
\if{latex}{\out{\hypertarget{method-getResponse}{}}}
\subsection{Method \code{getResponse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$getResponse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setResponse"></a>}}
\if{latex}{\out{\hypertarget{method-setResponse}{}}}
\subsection{Method \code{setResponse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$setResponse(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jsonReady"></a>}}
\if{latex}{\out{\hypertarget{method-jsonReady}{}}}
\subsection{Method \code{jsonReady()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$jsonReady()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interaction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
