% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionBuilder.R
\name{InteractionBuilder}
\alias{InteractionBuilder}
\title{Build an interaction and send it to the Ruby Standalone Mock Service}
\description{
Build an interaction and send it to the Ruby Standalone Mock Service

Build an interaction and send it to the Ruby Standalone Mock Service
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InteractionBuilder$new()}}
\item \href{#method-given}{\code{InteractionBuilder$given()}}
\item \href{#method-uponReceiving}{\code{InteractionBuilder$uponReceiving()}}
\item \href{#method-with}{\code{InteractionBuilder$with()}}
\item \href{#method-willRespondWith}{\code{InteractionBuilder$willRespondWith()}}
\item \href{#method-verify}{\code{InteractionBuilder$verify()}}
\item \href{#method-registerInteraction}{\code{InteractionBuilder$registerInteraction()}}
\item \href{#method-writePact}{\code{InteractionBuilder$writePact()}}
\item \href{#method-finalize}{\code{InteractionBuilder$finalize()}}
\item \href{#method-clone}{\code{InteractionBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$new(config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-given"></a>}}
\if{latex}{\out{\hypertarget{method-given}{}}}
\subsection{Method \code{given()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$given(providerState)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uponReceiving"></a>}}
\if{latex}{\out{\hypertarget{method-uponReceiving}{}}}
\subsection{Method \code{uponReceiving()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$uponReceiving(description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-with"></a>}}
\if{latex}{\out{\hypertarget{method-with}{}}}
\subsection{Method \code{with()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$with(request)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-willRespondWith"></a>}}
\if{latex}{\out{\hypertarget{method-willRespondWith}{}}}
\subsection{Method \code{willRespondWith()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$willRespondWith(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verify"></a>}}
\if{latex}{\out{\hypertarget{method-verify}{}}}
\subsection{Method \code{verify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$verify()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerInteraction"></a>}}
\if{latex}{\out{\hypertarget{method-registerInteraction}{}}}
\subsection{Method \code{registerInteraction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$registerInteraction(interaction)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writePact"></a>}}
\if{latex}{\out{\hypertarget{method-writePact}{}}}
\subsection{Method \code{writePact()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$writePact()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractionBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
