% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matcher.R
\name{Matcher}
\alias{Matcher}
\title{Create the matcher object}
\description{
This is used as a DSL for matchers
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-like}{\code{Matcher$like()}}
\item \href{#method-somethingLike}{\code{Matcher$somethingLike()}}
\item \href{#method-term}{\code{Matcher$term()}}
\item \href{#method-regex}{\code{Matcher$regex()}}
\item \href{#method-clone}{\code{Matcher$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-like"></a>}}
\if{latex}{\out{\hypertarget{method-like}{}}}
\subsection{Method \code{like()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Matcher$like(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-somethingLike"></a>}}
\if{latex}{\out{\hypertarget{method-somethingLike}{}}}
\subsection{Method \code{somethingLike()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Matcher$somethingLike(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-term"></a>}}
\if{latex}{\out{\hypertarget{method-term}{}}}
\subsection{Method \code{term()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Matcher$term(val, pattern)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regex"></a>}}
\if{latex}{\out{\hypertarget{method-regex}{}}}
\subsection{Method \code{regex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Matcher$regex(val, pattern)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Matcher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
