% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockServerConfig.R
\name{MockServerConfig}
\alias{MockServerConfig}
\title{Create the Mock Server configuration object}
\description{
Create the Mock Server configuration object

Create the Mock Server configuration object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MockServerConfig$new()}}
\item \href{#method-getHost}{\code{MockServerConfig$getHost()}}
\item \href{#method-setHost}{\code{MockServerConfig$setHost()}}
\item \href{#method-getPactSpecificationVersion}{\code{MockServerConfig$getPactSpecificationVersion()}}
\item \href{#method-setPactSpecificationVersion}{\code{MockServerConfig$setPactSpecificationVersion()}}
\item \href{#method-getPactLog}{\code{MockServerConfig$getPactLog()}}
\item \href{#method-setPactLog}{\code{MockServerConfig$setPactLog()}}
\item \href{#method-getPort}{\code{MockServerConfig$getPort()}}
\item \href{#method-setPort}{\code{MockServerConfig$setPort()}}
\item \href{#method-getConsumer}{\code{MockServerConfig$getConsumer()}}
\item \href{#method-setConsumer}{\code{MockServerConfig$setConsumer()}}
\item \href{#method-getProvider}{\code{MockServerConfig$getProvider()}}
\item \href{#method-setProvider}{\code{MockServerConfig$setProvider()}}
\item \href{#method-getPactDir}{\code{MockServerConfig$getPactDir()}}
\item \href{#method-setPactDir}{\code{MockServerConfig$setPactDir()}}
\item \href{#method-getPactFileWriteMode}{\code{MockServerConfig$getPactFileWriteMode()}}
\item \href{#method-setPactFileWriteMode}{\code{MockServerConfig$setPactFileWriteMode()}}
\item \href{#method-getUri}{\code{MockServerConfig$getUri()}}
\item \href{#method-hasCors}{\code{MockServerConfig$hasCors()}}
\item \href{#method-setCors}{\code{MockServerConfig$setCors()}}
\item \href{#method-isSecure}{\code{MockServerConfig$isSecure()}}
\item \href{#method-setSecure}{\code{MockServerConfig$setSecure()}}
\item \href{#method-clone}{\code{MockServerConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHost"></a>}}
\if{latex}{\out{\hypertarget{method-getHost}{}}}
\subsection{Method \code{getHost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getHost()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setHost"></a>}}
\if{latex}{\out{\hypertarget{method-setHost}{}}}
\subsection{Method \code{setHost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setHost(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPactSpecificationVersion"></a>}}
\if{latex}{\out{\hypertarget{method-getPactSpecificationVersion}{}}}
\subsection{Method \code{getPactSpecificationVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getPactSpecificationVersion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPactSpecificationVersion"></a>}}
\if{latex}{\out{\hypertarget{method-setPactSpecificationVersion}{}}}
\subsection{Method \code{setPactSpecificationVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setPactSpecificationVersion(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPactLog"></a>}}
\if{latex}{\out{\hypertarget{method-getPactLog}{}}}
\subsection{Method \code{getPactLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getPactLog()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPactLog"></a>}}
\if{latex}{\out{\hypertarget{method-setPactLog}{}}}
\subsection{Method \code{setPactLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setPactLog(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPort"></a>}}
\if{latex}{\out{\hypertarget{method-getPort}{}}}
\subsection{Method \code{getPort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getPort()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPort"></a>}}
\if{latex}{\out{\hypertarget{method-setPort}{}}}
\subsection{Method \code{setPort()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setPort(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConsumer"></a>}}
\if{latex}{\out{\hypertarget{method-getConsumer}{}}}
\subsection{Method \code{getConsumer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getConsumer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setConsumer"></a>}}
\if{latex}{\out{\hypertarget{method-setConsumer}{}}}
\subsection{Method \code{setConsumer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setConsumer(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProvider"></a>}}
\if{latex}{\out{\hypertarget{method-getProvider}{}}}
\subsection{Method \code{getProvider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getProvider()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProvider"></a>}}
\if{latex}{\out{\hypertarget{method-setProvider}{}}}
\subsection{Method \code{setProvider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setProvider(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPactDir"></a>}}
\if{latex}{\out{\hypertarget{method-getPactDir}{}}}
\subsection{Method \code{getPactDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getPactDir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPactDir"></a>}}
\if{latex}{\out{\hypertarget{method-setPactDir}{}}}
\subsection{Method \code{setPactDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setPactDir(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPactFileWriteMode"></a>}}
\if{latex}{\out{\hypertarget{method-getPactFileWriteMode}{}}}
\subsection{Method \code{getPactFileWriteMode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getPactFileWriteMode()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPactFileWriteMode"></a>}}
\if{latex}{\out{\hypertarget{method-setPactFileWriteMode}{}}}
\subsection{Method \code{setPactFileWriteMode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setPactFileWriteMode(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUri"></a>}}
\if{latex}{\out{\hypertarget{method-getUri}{}}}
\subsection{Method \code{getUri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$getUri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasCors"></a>}}
\if{latex}{\out{\hypertarget{method-hasCors}{}}}
\subsection{Method \code{hasCors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$hasCors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCors"></a>}}
\if{latex}{\out{\hypertarget{method-setCors}{}}}
\subsection{Method \code{setCors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setCors(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isSecure"></a>}}
\if{latex}{\out{\hypertarget{method-isSecure}{}}}
\subsection{Method \code{isSecure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$isSecure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setSecure"></a>}}
\if{latex}{\out{\hypertarget{method-setSecure}{}}}
\subsection{Method \code{setSecure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$setSecure(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
