% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockServerHttpService.R
\name{MockServerHttpService}
\alias{MockServerHttpService}
\title{Http Service that interacts with the Ruby Standalone Mock Server}
\description{
Http Service that interacts with the Ruby Standalone Mock Server

Http Service that interacts with the Ruby Standalone Mock Server
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MockServerHttpService$new()}}
\item \href{#method-healthCheck}{\code{MockServerHttpService$healthCheck()}}
\item \href{#method-deleteAllInteractions}{\code{MockServerHttpService$deleteAllInteractions()}}
\item \href{#method-registerInteraction}{\code{MockServerHttpService$registerInteraction()}}
\item \href{#method-verifyInteractions}{\code{MockServerHttpService$verifyInteractions()}}
\item \href{#method-getPactJson}{\code{MockServerHttpService$getPactJson()}}
\item \href{#method-performPactConsumerHttpRequest}{\code{MockServerHttpService$performPactConsumerHttpRequest()}}
\item \href{#method-clone}{\code{MockServerHttpService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$new(config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-healthCheck"></a>}}
\if{latex}{\out{\hypertarget{method-healthCheck}{}}}
\subsection{Method \code{healthCheck()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$healthCheck()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deleteAllInteractions"></a>}}
\if{latex}{\out{\hypertarget{method-deleteAllInteractions}{}}}
\subsection{Method \code{deleteAllInteractions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$deleteAllInteractions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-registerInteraction"></a>}}
\if{latex}{\out{\hypertarget{method-registerInteraction}{}}}
\subsection{Method \code{registerInteraction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$registerInteraction(interaction)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verifyInteractions"></a>}}
\if{latex}{\out{\hypertarget{method-verifyInteractions}{}}}
\subsection{Method \code{verifyInteractions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$verifyInteractions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPactJson"></a>}}
\if{latex}{\out{\hypertarget{method-getPactJson}{}}}
\subsection{Method \code{getPactJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$getPactJson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-performPactConsumerHttpRequest"></a>}}
\if{latex}{\out{\hypertarget{method-performPactConsumerHttpRequest}{}}}
\subsection{Method \code{performPactConsumerHttpRequest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$performPactConsumerHttpRequest(baseUrl, consumerRequest)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MockServerHttpService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
