% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_cite.R
\name{pact_cite}
\alias{pact_cite}
\title{Cite an output/asset available from the Pandemic PACT Figshare repository}
\usage{
pact_cite(pact_client, id)
}
\arguments{
\item{pact_client}{An interface client to the Pandemic PACT Figshare
repository. This is usually set/created through a call to
\code{pact_client_set()}.}

\item{id}{A unique integer value identifying a specific file in the
repository.}
}
\value{
A string for recommended bibliographic citation for specific
Pandemic PACT output/asset in DateCite bibliographic style.
}
\description{
Cite an output/asset available from the Pandemic PACT Figshare repository
}
\examples{
pact_client <- pact_client_set()
pact_cite(pact_client, id = 24763548)

}
