% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_client.R
\name{pact_client_set}
\alias{pact_client_set}
\title{Initiate a new Pandemic PACT Figshare client session}
\usage{
pact_client_set()
}
\value{
A local client to the Figahre service.
}
\description{
Initiate a new Pandemic PACT Figshare client session
}
\examples{
pact_client_set()

}
