% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_list.R
\name{pact_figshare_list}
\alias{pact_figshare_list}
\title{List all available outputs/assets from Pandemic PACT's Figshare repositories.}
\usage{
pact_figshare_list(pact_client)
}
\arguments{
\item{pact_client}{An interface client to the Pandemic PACT Figshare
repository. This is usually set/created through a call to
\code{pact_client_set()}.}
}
\value{
A \code{tibble::tibble()} of available outputs/assets from the
Pandemic PACT Figshare repository.
}
\description{
List all available outputs/assets from Pandemic PACT's Figshare repositories.
}
\examples{
\dontrun{
  pact_figshare_list(pact_client = pact_client_set())
}

}
