% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pact_mpox_priority}
\alias{pact_mpox_priority}
\title{Pandemic PACT Mpox Priorities}
\format{
A data.frame with 4 columns and 23 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{mpox_priority_code} \tab Two digit Mpox priority code \cr
   \emph{mpox_priority} \tab Mpox priority name \cr
   \emph{mpox_subpriority_code} \tab Four digit Mpox sub-priority code \cr
   \emph{mpox_subpriority} \tab Mpox sub-priority name \cr
}
}
\source{
Pandemic PACT website https://www.pandemicpact.org/
}
\usage{
pact_mpox_priority
}
\description{
Pandemic PACT Mpox Priorities
}
\examples{
pact_mpox_priority

}
\keyword{datasets}
