% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_process.R
\name{pact_process_data_tracker}
\alias{pact_process_data_tracker}
\title{Process Pandemic PACT data}
\usage{
pact_process_data_tracker(df)
}
\arguments{
\item{df}{A data.frame of the Pandemic PACT dataset from the Figshare
repository}
}
\value{
A tibble
}
\description{
Process Pandemic PACT data
}
\examples{
\dontrun{
  pact_data <- pact_read_data_tracker(pact_client_set())
  pact_process_data_tracker(df = pact_data)
}

}
