% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_process_figshare.R
\name{pact_process_figshare}
\alias{pact_process_figshare}
\alias{pact_process_figshare_category}
\alias{pact_process_figshare_category_pathogen}
\alias{pact_process_figshare_category_funder}
\title{Process Pandemic PACT data}
\usage{
pact_process_figshare(df)

pact_process_figshare_category(df, category, other = NULL, nest = FALSE)

pact_process_figshare_category_pathogen(df)

pact_process_figshare_category_funder(df)
}
\arguments{
\item{df}{A data.frame of the Pandemic PACT dataset from the Figshare
repository or from the website.}

\item{category}{A character value for the variable category to look for
in the fields for \code{df} for collapsing multiple fields into one.}

\item{other}{A character value for the name of the variable in \code{df} for
values for other for fields that have an other option. Default to NULL to
indicate that field \code{category} has no other option.}

\item{nest}{Logical. Should variable/fields with multiple values be nested?
Default to FALSE.}
}
\value{
A tibble of processed Pandemic PACT dataset.
}
\description{
Process Pandemic PACT data
}
\examples{
\dontrun{
  pact_data <- pact_read_figshare(pact_client_set())
  pact_process_figshare(df = pact_data)
}

}
