% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_read.R
\name{pact_read_data_figshare}
\alias{pact_read_data_figshare}
\alias{pact_read_data_dictionary}
\alias{pact_read_data_website}
\title{Read datasets from the Pandemic PACT Figshare repository}
\usage{
pact_read_data_figshare(pact_client, tracker_type = c("labelled", "raw"))

pact_read_data_dictionary(pact_client)

pact_read_data_website(nest = TRUE, col_list = TRUE)
}
\arguments{
\item{pact_client}{An interface client to the Pandemic PACT Figshare
repository. This is usually set/created through a call to
\code{pact_client_set()}.}

\item{tracker_type}{Either "labelled" or "raw". Default is "labelled".}

\item{nest}{Logical. Should variable/fields with multiple values be nested?
Default to FALSE.}

\item{col_list}{Logical. Should variable/fields with multiple values be
made into column lists? Only evaluated if \code{nest = TRUE}.}
}
\value{
A data.frame of the requested dataset. For
\code{pact_read_data_website()}, a tibble of the dataset from the website
structured based on \code{nest} and \code{col_list} specifications.
}
\description{
Read datasets from the Pandemic PACT Figshare repository
}
\examples{
\dontrun{
  pact_client <- pact_client_set()
  pact_read_data_dictionary(pact_client)
}

}
