% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_read_figshare.R
\name{pact_data_read_figshare}
\alias{pact_data_read_figshare}
\alias{pact_read_figshare_dictionary}
\title{Read Pandemic PACT tracker dataset from the Pandemic PACT Figshare repository}
\usage{
pact_data_read_figshare(pact_client, latest = TRUE, version = NULL)

pact_read_figshare_dictionary(pact_client)
}
\arguments{
\item{pact_client}{An interface client to the Pandemic PACT Figshare
repository. This is usually set/created through a call to
\code{pact_client_set()}.}

\item{latest}{Logical. If \code{TRUE} (default), the latest version of the
Pandemic PACT dataset is read.}

\item{version}{A character string specifying the version of the Pandemic PACT
dataset to read. This should be in the format "YYYY-MM-DD". If \code{latest} is
set to \code{TRUE}, this argument is ignored. If \code{version} is specified, it must
match one of the available versions in the Pandemic PACT Figshare repository.
If it does not match, a warning will be issued and the latest version will
be read instead. If \code{version} is not specified and \code{latest} is set to
\code{FALSE}, a warning will be issued and the latest version will be read
instead.}
}
\value{
A data.frame of the requested dataset.
}
\description{
Read Pandemic PACT tracker dataset from the Pandemic PACT Figshare repository
}
\examples{
\dontrun{
  pact_client <- pact_client_set()
  pact_read_figshare_dictionary(pact_client)
}

}
