% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_read_website.R
\name{pact_read_website}
\alias{pact_read_website}
\title{Read datasets from the Pandemic PACT website}
\usage{
pact_read_website(nest = TRUE, col_list = TRUE)
}
\arguments{
\item{nest}{Logical. Should variable/fields with multiple values be nested?
Default to FALSE.}

\item{col_list}{Logical. Should variable/fields with multiple values be
made into column lists? Only evaluated if \code{nest = TRUE}.}
}
\value{
A tibble of the dataset from the website structured based on
\code{nest} and \code{col_list} specifications.
}
\description{
Read datasets from the Pandemic PACT website
}
\examples{
\dontrun{
  pact_read_website(pact_client)
}

}
