% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pact_research_category}
\alias{pact_research_category}
\title{Pandemic PACT Research Categories}
\format{
A data.frame with 4 columns and 70 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{research_category_code} \tab Two digit research category code \cr
   \emph{research_category} \tab Research category name \cr
   \emph{research_subcategory_code} \tab Four digit research sub-category code \cr
   \emph{research_subcategory} \tab Research sub-category name \cr
}
}
\source{
Pandemic PACT website https://www.pandemicpact.org/
}
\usage{
pact_research_category
}
\description{
Pandemic PACT Research Categories
}
\examples{
pact_research_category

}
\keyword{datasets}
