% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pactr-package.R
\docType{package}
\name{pactr}
\alias{pactr-package}
\alias{pactr}
\title{An Interface to the Pandemic PACT Database}
\description{
The Pandemic PACT \url{https://www.pandemicpact.org/} monitors and analyses
global funding and research evidence related to diseases with pandemic
potential, as well as broader research preparedness efforts, and is equipped
to pivot in response to outbreaks. It collects, curates, codes, and analyses
data in alignment with WHO priority diseases and other selected illnesses,
including pandemic influenza, mpox, and plague. Pandemic PACT aims to guide
policy and decision-making for research funders, policymakers, researchers,
multilateral agencies. The database is publicly available for download from
its website and from Figshare. This package interfaces with the project's
Figshare repository to provide programmatic access to the database along with
other data products provided by the project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OxfordIHTM/pactr}
  \item \url{https://oxford-ihtm.io/pactr/}
  \item \url{http://oxford-ihtm.io/pactr/}
  \item Report bugs at \url{https://github.com/OxfordIHTM/pactr/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
