% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PactScripts.R
\name{PactScripts}
\alias{PactScripts}
\title{Wrapper for all the Ruby Pact Scripts to call}
\description{
Wrapper for all the Ruby Pact Scripts to call

Wrapper for all the Ruby Pact Scripts to call
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PactScripts$new()}}
\item \href{#method-clone}{\code{PactScripts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactScripts$new(mockService, stubService, providerVerifier, pactMessage)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactScripts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
