% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PactUtility.R
\name{PactUtility}
\alias{PactUtility}
\title{Encapsulate some helper functions}
\description{
Encapsulate some helper functions

Encapsulate some helper functions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckR6ClassType}{\code{PactUtility$CheckR6ClassType()}}
\item \href{#method-EnforceR6ClassType}{\code{PactUtility$EnforceR6ClassType()}}
\item \href{#method-clone}{\code{PactUtility$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckR6ClassType"></a>}}
\if{latex}{\out{\hypertarget{method-CheckR6ClassType}{}}}
\subsection{Method \code{CheckR6ClassType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactUtility$CheckR6ClassType(r6Obj, className)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnforceR6ClassType"></a>}}
\if{latex}{\out{\hypertarget{method-EnforceR6ClassType}{}}}
\subsection{Method \code{EnforceR6ClassType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactUtility$EnforceR6ClassType(r6Obj, className)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactUtility$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
