% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PactVerifier.R
\name{PactVerifier}
\alias{PactVerifier}
\title{Create the Verifier object}
\description{
Wrapper for the Ruby Standalone Verifier service.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PactVerifier$new()}}
\item \href{#method-getArguments}{\code{PactVerifier$getArguments()}}
\item \href{#method-verify}{\code{PactVerifier$verify()}}
\item \href{#method-verifyFiles}{\code{PactVerifier$verifyFiles()}}
\item \href{#method-verifyAll}{\code{PactVerifier$verifyAll()}}
\item \href{#method-verifyAllForTag}{\code{PactVerifier$verifyAllForTag()}}
\item \href{#method-verifyAction}{\code{PactVerifier$verifyAction()}}
\item \href{#method-clone}{\code{PactVerifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$new(config, install_pact_dir = "../pact-bin")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getArguments"></a>}}
\if{latex}{\out{\hypertarget{method-getArguments}{}}}
\subsection{Method \code{getArguments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$getArguments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verify"></a>}}
\if{latex}{\out{\hypertarget{method-verify}{}}}
\subsection{Method \code{verify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$verify(
  consumer_name,
  tag = NA,
  consumer_version = NA,
  process_x_args = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verifyFiles"></a>}}
\if{latex}{\out{\hypertarget{method-verifyFiles}{}}}
\subsection{Method \code{verifyFiles()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$verifyFiles(files, process_x_args = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verifyAll"></a>}}
\if{latex}{\out{\hypertarget{method-verifyAll}{}}}
\subsection{Method \code{verifyAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$verifyAll(process_x_args = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verifyAllForTag"></a>}}
\if{latex}{\out{\hypertarget{method-verifyAllForTag}{}}}
\subsection{Method \code{verifyAllForTag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$verifyAllForTag(tag, process_x_args = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-verifyAction"></a>}}
\if{latex}{\out{\hypertarget{method-verifyAction}{}}}
\subsection{Method \code{verifyAction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$verifyAction(args, process_x_args = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
