% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PactVerifierConfig.R
\name{PactVerifierConfig}
\alias{PactVerifierConfig}
\title{Verifier Configuation object to configure use of the Pact Verifier}
\description{
Verifier Configuation object to configure use of the Pact Verifier

Verifier Configuation object to configure use of the Pact Verifier
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PactVerifierConfig$new()}}
\item \href{#method-getProviderBaseUrl}{\code{PactVerifierConfig$getProviderBaseUrl()}}
\item \href{#method-setProviderBaseUrl}{\code{PactVerifierConfig$setProviderBaseUrl()}}
\item \href{#method-getProviderStatesSetupUrl}{\code{PactVerifierConfig$getProviderStatesSetupUrl()}}
\item \href{#method-setProviderStatesSetupUrl}{\code{PactVerifierConfig$setProviderStatesSetupUrl()}}
\item \href{#method-getProviderName}{\code{PactVerifierConfig$getProviderName()}}
\item \href{#method-setProviderName}{\code{PactVerifierConfig$setProviderName()}}
\item \href{#method-getProviderVersion}{\code{PactVerifierConfig$getProviderVersion()}}
\item \href{#method-setProviderVersion}{\code{PactVerifierConfig$setProviderVersion()}}
\item \href{#method-getPublishResults}{\code{PactVerifierConfig$getPublishResults()}}
\item \href{#method-setPublishResults}{\code{PactVerifierConfig$setPublishResults()}}
\item \href{#method-getBrokerUri}{\code{PactVerifierConfig$getBrokerUri()}}
\item \href{#method-setBrokerUri}{\code{PactVerifierConfig$setBrokerUri()}}
\item \href{#method-getBrokerUserName}{\code{PactVerifierConfig$getBrokerUserName()}}
\item \href{#method-setBrokerUser}{\code{PactVerifierConfig$setBrokerUser()}}
\item \href{#method-getBrokerPassword}{\code{PactVerifierConfig$getBrokerPassword()}}
\item \href{#method-setBrokerPassword}{\code{PactVerifierConfig$setBrokerPassword()}}
\item \href{#method-getCustomProviderHeaders}{\code{PactVerifierConfig$getCustomProviderHeaders()}}
\item \href{#method-setCustomProviderHeaders}{\code{PactVerifierConfig$setCustomProviderHeaders()}}
\item \href{#method-getVerbose}{\code{PactVerifierConfig$getVerbose()}}
\item \href{#method-setVerbose}{\code{PactVerifierConfig$setVerbose()}}
\item \href{#method-getFormat}{\code{PactVerifierConfig$getFormat()}}
\item \href{#method-setFormat}{\code{PactVerifierConfig$setFormat()}}
\item \href{#method-getProcessTimeout}{\code{PactVerifierConfig$getProcessTimeout()}}
\item \href{#method-setProcessTimeout}{\code{PactVerifierConfig$setProcessTimeout()}}
\item \href{#method-getProcessIdleTimeout}{\code{PactVerifierConfig$getProcessIdleTimeout()}}
\item \href{#method-setProcessIdleTimeout}{\code{PactVerifierConfig$setProcessIdleTimeout()}}
\item \href{#method-clone}{\code{PactVerifierConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProviderBaseUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getProviderBaseUrl}{}}}
\subsection{Method \code{getProviderBaseUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getProviderBaseUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProviderBaseUrl"></a>}}
\if{latex}{\out{\hypertarget{method-setProviderBaseUrl}{}}}
\subsection{Method \code{setProviderBaseUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setProviderBaseUrl(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProviderStatesSetupUrl"></a>}}
\if{latex}{\out{\hypertarget{method-getProviderStatesSetupUrl}{}}}
\subsection{Method \code{getProviderStatesSetupUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getProviderStatesSetupUrl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProviderStatesSetupUrl"></a>}}
\if{latex}{\out{\hypertarget{method-setProviderStatesSetupUrl}{}}}
\subsection{Method \code{setProviderStatesSetupUrl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setProviderStatesSetupUrl(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProviderName"></a>}}
\if{latex}{\out{\hypertarget{method-getProviderName}{}}}
\subsection{Method \code{getProviderName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getProviderName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProviderName"></a>}}
\if{latex}{\out{\hypertarget{method-setProviderName}{}}}
\subsection{Method \code{setProviderName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setProviderName(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProviderVersion"></a>}}
\if{latex}{\out{\hypertarget{method-getProviderVersion}{}}}
\subsection{Method \code{getProviderVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getProviderVersion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProviderVersion"></a>}}
\if{latex}{\out{\hypertarget{method-setProviderVersion}{}}}
\subsection{Method \code{setProviderVersion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setProviderVersion(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPublishResults"></a>}}
\if{latex}{\out{\hypertarget{method-getPublishResults}{}}}
\subsection{Method \code{getPublishResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getPublishResults()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPublishResults"></a>}}
\if{latex}{\out{\hypertarget{method-setPublishResults}{}}}
\subsection{Method \code{setPublishResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setPublishResults(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBrokerUri"></a>}}
\if{latex}{\out{\hypertarget{method-getBrokerUri}{}}}
\subsection{Method \code{getBrokerUri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getBrokerUri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBrokerUri"></a>}}
\if{latex}{\out{\hypertarget{method-setBrokerUri}{}}}
\subsection{Method \code{setBrokerUri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setBrokerUri(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBrokerUserName"></a>}}
\if{latex}{\out{\hypertarget{method-getBrokerUserName}{}}}
\subsection{Method \code{getBrokerUserName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getBrokerUserName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBrokerUser"></a>}}
\if{latex}{\out{\hypertarget{method-setBrokerUser}{}}}
\subsection{Method \code{setBrokerUser()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setBrokerUser(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBrokerPassword"></a>}}
\if{latex}{\out{\hypertarget{method-getBrokerPassword}{}}}
\subsection{Method \code{getBrokerPassword()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getBrokerPassword()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBrokerPassword"></a>}}
\if{latex}{\out{\hypertarget{method-setBrokerPassword}{}}}
\subsection{Method \code{setBrokerPassword()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setBrokerPassword(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCustomProviderHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-getCustomProviderHeaders}{}}}
\subsection{Method \code{getCustomProviderHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getCustomProviderHeaders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCustomProviderHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-setCustomProviderHeaders}{}}}
\subsection{Method \code{setCustomProviderHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setCustomProviderHeaders(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getVerbose"></a>}}
\if{latex}{\out{\hypertarget{method-getVerbose}{}}}
\subsection{Method \code{getVerbose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getVerbose()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setVerbose"></a>}}
\if{latex}{\out{\hypertarget{method-setVerbose}{}}}
\subsection{Method \code{setVerbose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setVerbose(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFormat"></a>}}
\if{latex}{\out{\hypertarget{method-getFormat}{}}}
\subsection{Method \code{getFormat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getFormat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setFormat"></a>}}
\if{latex}{\out{\hypertarget{method-setFormat}{}}}
\subsection{Method \code{setFormat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setFormat(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcessTimeout"></a>}}
\if{latex}{\out{\hypertarget{method-getProcessTimeout}{}}}
\subsection{Method \code{getProcessTimeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getProcessTimeout()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProcessTimeout"></a>}}
\if{latex}{\out{\hypertarget{method-setProcessTimeout}{}}}
\subsection{Method \code{setProcessTimeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setProcessTimeout(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcessIdleTimeout"></a>}}
\if{latex}{\out{\hypertarget{method-getProcessIdleTimeout}{}}}
\subsection{Method \code{getProcessIdleTimeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$getProcessIdleTimeout()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setProcessIdleTimeout"></a>}}
\if{latex}{\out{\hypertarget{method-setProcessIdleTimeout}{}}}
\subsection{Method \code{setProcessIdleTimeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$setProcessIdleTimeout(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PactVerifierConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
