% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_author_names}
\alias{process_author_names}
\title{Process author names}
\usage{
process_author_names(authors)
}
\arguments{
\item{authors}{A character vector of author names}
}
\value{
A concatenated string of author names formatted as per person
specification.
}
\description{
Process author names
}
\examples{
\dontrun{
  process_author_names(c("John Doe", "Jane Doe"))
}

}
\keyword{internal}
