% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProviderResponse.R
\name{ProviderResponse}
\alias{ProviderResponse}
\title{Create the Provider Response object}
\description{
This is used as a DSL representing a response
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProviderResponse$new()}}
\item \href{#method-getStatus}{\code{ProviderResponse$getStatus()}}
\item \href{#method-setStatus}{\code{ProviderResponse$setStatus()}}
\item \href{#method-getHeaders}{\code{ProviderResponse$getHeaders()}}
\item \href{#method-setHeaders}{\code{ProviderResponse$setHeaders()}}
\item \href{#method-getBody}{\code{ProviderResponse$getBody()}}
\item \href{#method-setBody}{\code{ProviderResponse$setBody()}}
\item \href{#method-jsonReady}{\code{ProviderResponse$jsonReady()}}
\item \href{#method-clone}{\code{ProviderResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatus"></a>}}
\if{latex}{\out{\hypertarget{method-getStatus}{}}}
\subsection{Method \code{getStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$getStatus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setStatus"></a>}}
\if{latex}{\out{\hypertarget{method-setStatus}{}}}
\subsection{Method \code{setStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$setStatus(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-getHeaders}{}}}
\subsection{Method \code{getHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$getHeaders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setHeaders"></a>}}
\if{latex}{\out{\hypertarget{method-setHeaders}{}}}
\subsection{Method \code{setHeaders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$setHeaders(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBody"></a>}}
\if{latex}{\out{\hypertarget{method-getBody}{}}}
\subsection{Method \code{getBody()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$getBody()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBody"></a>}}
\if{latex}{\out{\hypertarget{method-setBody}{}}}
\subsection{Method \code{setBody()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$setBody(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jsonReady"></a>}}
\if{latex}{\out{\hypertarget{method-jsonReady}{}}}
\subsection{Method \code{jsonReady()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$jsonReady()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProviderResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
