% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_country_info}
\alias{who_country_info}
\title{World Health Organization (WHO) Country Information}
\format{
A data frame with 5 columns and 194 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{country_iso3c} \tab ISO3 country code \cr
   \emph{who_short_name} \tab Country short name given by WHO \cr
   \emph{formal_name} \tab Country formal name \cr
   \emph{who_region} \tab WHO region name to which country belongs to \cr
   \emph{un_region} \tab UN region name to which country belongs to \cr
}
}
\source{
Data are drawn from WHO's Data Country list
https://data.who.int/countries/
}
\usage{
who_country_info
}
\description{
World Health Organization (WHO) Country Information
}
\examples{
who_country_info

}
\keyword{datasets}
