% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_who_region}
\alias{get_who_region}
\alias{get_who_regions}
\title{Get WHO region}
\usage{
get_who_region(x)

get_who_regions(x)
}
\arguments{
\item{x}{A character value or vector of values of WHO countries to get the
WHO region/s of.}
}
\value{
A character value or vector of values of WHO regions.
}
\description{
Get WHO region
}
\examples{
\dontrun{
  get_who_regions("United Kingdom")
}

}
