% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_download.R
\name{pact_download_figshare}
\alias{pact_download_figshare}
\alias{pact_download_website}
\title{Download outputs/assets from Pandemic PACT Figshare repository}
\usage{
pact_download_figshare(pact_client, id, path, overwrite = FALSE, quiet = TRUE)

pact_download_website(path, overwrite = FALSE, quiet = TRUE)
}
\arguments{
\item{pact_client}{An interface client to the Pandemic PACT Figshare
repository. This is usually set/created through a call to
\code{pact_client_set()}.}

\item{id}{A unique integer value identifying a specific file in the
repository.}

\item{path}{The local directory where file is to be downloaded.}

\item{overwrite}{Logical. Should existing files be overwritten? If TRUE,
existing files will be overwritten. Default is FALSE.}

\item{quiet}{Logical. If TRUE (default), download progress is not displayed.}
}
\value{
The full path of the downloaded file.
}
\description{
Download outputs/assets from Pandemic PACT Figshare repository
}
\examples{
\dontrun{
  ## From Figshare
  pact_client <- pact_client_set()
  pact_download_figshare(pact_client, id = 25827649, path = tempdir())

  ## From website
  pact_download_website(path = tempdir())
}


}
