% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_process.R
\name{pact_process_topic_group}
\alias{pact_process_topic_group}
\alias{pact_process_disease}
\alias{pact_process_category}
\title{Process variable of interest from Pandemic PACT website data by a grouping
variable}
\usage{
pact_process_topic_group(
  df,
  topic,
  group = NULL,
  outcome = c("frequency", "money")
)

pact_process_disease(df, group = NULL, outcome = c("frequency", "money"))

pact_process_category(df, group = NULL, outcome = c("frequency", "money"))
}
\arguments{
\item{df}{A data.frame of the Pandemic PACT dataset from the Figshare
repository.}

\item{topic}{A character value of the variable name in \code{df} for the topic of
interest.}

\item{group}{A character value or vector of up to two values of the variable
name/s  in \code{df} to use as grouping variable/s. When specified as NULL
(default), no grouping is applied to come up with value of \code{outcome}
based on the \code{topic} of interest. When specified as NULL (default), no
grouping is applied.}

\item{outcome}{The type of outcome. Either \emph{"frequency"} or \emph{"money"}.
Default is \emph{"frequency"}.}
}
\value{
A data.frame structured based on specification. If \code{group} is NULL,
the data.frame presents values for \code{topic} as first column and then either
frequencies/counts of grants per \code{topic} value or sum of monetary amount of
grants per \code{topic}. if \code{group} has one value, the data.frame presents
values for \code{group} as first column followed by either frequencies/counts
of grants per \code{group} or sum of monetary amount of grants per \code{group} then
followed by the \code{topics} within each \code{group} followed by either
frequencies/counts of grants per \code{topic} by \code{group} or sum of monetary
amount of grants per \code{topic} by \code{group}.
}
\description{
Process variable of interest from Pandemic PACT website data by a grouping
variable
}
\examples{
\dontrun{
  df <- pact_read_website()
  pact_process_topic_group(df, topic = "Disease")
}

}
