% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_process.R
\name{pact_process_topic_group}
\alias{pact_process_topic_group}
\alias{pact_process_disease}
\alias{pact_process_category}
\title{Process variable of interest from Pandemic PACT website data by a grouping
variable}
\usage{
pact_process_topic_group(
  pact_data_list_cols,
  topic,
  group = NULL,
  outcome = c("frequency", "money"),
  na_values = NULL
)

pact_process_disease(
  pact_data_list_cols,
  group = NULL,
  outcome = c("frequency", "money")
)

pact_process_category(
  pact_data_list_cols,
  topic = c("ResearchCat", "ResearchSubcat"),
  split_grants = TRUE,
  outcome = c("frequency", "money")
)
}
\arguments{
\item{pact_data_list_cols}{A data.frame for the Pandemic PACT dataset read
from the Pandemic PACT website that has already been pre-processed to have
list columns for nested variables. This is usually obtained via a call to
\code{pact_process_website()} with \code{col_list = TRUE}.}

\item{topic}{A character value of the variable name in \code{pact_data} for the
topic of interest.}

\item{group}{A character value or vector of up to two values of the variable
name/s  in \code{pact_data} to use as grouping variable/s. When specified as
NULL (default), no grouping is applied to come up with value of \code{outcome}
based on the \code{topic} of interest.}

\item{outcome}{The type of outcome. Either \emph{"frequency"} or \emph{"money"}.
Default is \emph{"frequency"}.}

\item{na_values}{A character value or vector of values for strings to be
considered as NA for \code{topic} and \code{group}. If NULL (default), \code{topic} and
\code{group} are kept as is.}

\item{split_grants}{Logical. Should grants be divided into those with
specified or unspecified amount of funding. Default to TRUE.}
}
\value{
A data.frame structured based on specification. If \code{group} is NULL,
the data.frame presents values for \code{topic} as first column and then either
frequencies/counts of grants per \code{topic} value or sum of monetary amount of
grants per \code{topic}. if \code{group} has one value, the data.frame presents
values for \code{group} as first column followed by either frequencies/counts
of grants per \code{group} or sum of monetary amount of grants per \code{group} then
followed by the \code{topics} within each \code{group} followed by either
frequencies/counts of grants per \code{topic} by \code{group} or sum of monetary
amount of grants per \code{topic} by \code{group}.
}
\description{
Process variable of interest from Pandemic PACT website data by a grouping
variable
}
\examples{
\dontrun{
  df <- pact_read_website() |> pact_process_website()
  pact_process_topic_group(df, topic = "Disease")
}

}
