% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_read.R
\name{pact_read_data_tracker}
\alias{pact_read_data_tracker}
\alias{pact_read_data_dictionary}
\title{Read datasets from the Pandemic PACT Figshare repository}
\usage{
pact_read_data_tracker(pact_client, tracker_type = c("labelled", "raw"))

pact_read_data_dictionary(pact_client)
}
\arguments{
\item{pact_client}{An interface client to the Pandemic PACT Figshare
repository. This is usually set/created through a call to
\code{pact_client_set()}.}

\item{tracker_type}{Either "labelled" or "raw". Default is "labelled".}
}
\value{
A data.frame of the requested dataset.
}
\description{
Read datasets from the Pandemic PACT Figshare repository
}
\examples{
pact_client <- pact_client_set()
pact_read_data_dictionary(pact_client)

}
