% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_read_website.R
\name{pact_read_website}
\alias{pact_read_website}
\title{Read datasets from the Pandemic PACT website}
\usage{
pact_read_website(.url = NULL)
}
\arguments{
\item{.url}{The URL for the website dataset.}
}
\value{
A tibble of the Pandemic PACT dataset from the website.
}
\description{
Read datasets from the Pandemic PACT website
}
\examples{
\dontrun{
  pact_read_website()
}

}
